move estiamte stabilization to estimate registration
